Rework the change set processing.  These changes:
1) Enable multi-threaded processing of the changes for certain types of
change sets.

2) Enable updating the the change set xml document in-place - rather
than getting the xml, deserializing, updating, serializing, and storing
the document for each change in the change set.  This helps performance
a little bit there still is room for improvement here based on
conversations on the existdb mailing list.

3) Enable a 'bulk load mode' for processing a complete changeset xml
document.  During bulk load mode, all updates to the change set document
that will be stored in the DB are held in memory until the entire change
set is processed - and then written once at the end.  This (combined
with 1) give a massive improvement in performance compared to 2).  This
prevents existDB from thrashing about while indexing and deleting the
same content over and over again.