This commit does the upgrade to existDB 2.1.
There are still some library issues to work out - currently, this only
runs within OSGI if the XML libraries are updated in the
jre/lib/extensions folder.

The necessary libraries are included in the zip file
'endorsed-for-jre-lib-folders.zip'

There are a few TODOs in this code that I still need to discuss with
Kevin.

The required configuration for existDB has changed - see the included
suggested-conf.xml for the current suggested configuration.